requirements.txt
torch>=1.12.0
torchvision
transformers
diffusers>=0.21.0
accelerate
openai
openai-clip
datasets
peft
safetensors
Pillow
huggingface-hub
deep_translator
wandb
spacy